Made it where Anys are copyable.
Also added some tests to show where we are heading with being able to
use diagrams in charmapper.  Kind of got stuck because we really need to
have some kind of stack frame for diagram evaluation so that we're not
actually modifying the diagram nodes when we do an evaluation of the
diagram.