Working on worldwrapper_test
Working on adding testSettingStateWithSceneWithBody().  I realized that
the wrapper structure isn't right.  Assuming the tree actually represents
the state of the world, we have to be showing the background frame of
a scene, and we have to be showing the body positions as indices into the
frame instead of the actual frame value.