Work on support for object members in expressions.
I made it where objects have functions which let us generate members.  This
lets the std::function handle the type erasure, so we can easily pass objects
around.  This doesn't really work well when we need to make sure we have
a particular kind of object that comes out of an expression though.  For
example, if we have a PosExpr, we'll need to extract the actual BodyLink
out of it.