Made it where labels can't be changed again.
There is are now separate Wrapper::tag() and Wrapper::label() methods.  The
label can change, but not directly.  If a mechanism needs to be provided for
having a descriptive label for item and being able to change it, then
something else will have to be done, such as providing a name child.